refactor(storage): move tokenize/tokenOverlapScore into search module
These are keyword search internals, not core storage utilities. Also
removes StorageSearchResult from the public storage barrel export.
