+++ strands-ts/src/storage/search/keyword.ts
@@ -1,7 +1,24 @@
 import type { Storage, StorageSearchResult } from '../storage.js'
 import type { SearchStrategy } from './types.js'
 
-import { tokenize, tokenOverlapScore } from '../storage.js'
+/** Split text into a set of lowercased word tokens for keyword matching. */
+export function tokenize(text: string): Set<string> {
+  return new Set(
+    text
+      .toLowerCase()
+      .split(/[^\p{L}\p{N}_]+/u)
+      .filter(Boolean)
+  )
+}
+
+/** Lexical relevance score: distinct query tokens that appear in the content. Returns 0 when there is no overlap. */
+export function tokenOverlapScore(queryTokens: Set<string>, content: string): number {
+  let score = 0
+  for (const token of tokenize(content)) {
+    if (queryTokens.has(token)) score++
+  }
+  return score
+}
 
 /**
  * Keyword search strategy using token-overlap scoring.
